[ast][mir] Combile MIR AST node files
