+++ samlang-core-optimization/__tests__/mir-common-subexpression-elimination-optimization.test.ts
@@ -13,8 +13,9 @@ import {
   MIR_FUNCTION_CALL,
   MIR_IF_ELSE,
   MIR_CAST,
-} from 'samlang-core-ast/mir-expressions';
-import { MIR_BOOL_TYPE, MIR_INT_TYPE } from 'samlang-core-ast/mir-types';
+  MIR_BOOL_TYPE,
+  MIR_INT_TYPE,
+} from 'samlang-core-ast/mir-nodes';
 
 const assertCorrectlyOptimized = (statements: MidIRStatement[], expected: string): void => {
   expect(
