+++ samlang-core-optimization/__tests__/mir-inline-optimization.test.ts
@@ -19,9 +19,12 @@ import {
   MIR_CAST,
   MIR_STRUCT_INITIALIZATION,
   MidIRStatement,
-} from 'samlang-core-ast/mir-expressions';
-import { debugPrintMidIRFunction, MidIRFunction } from 'samlang-core-ast/mir-toplevel';
-import { MIR_BOOL_TYPE, MIR_FUNCTION_TYPE, MIR_INT_TYPE } from 'samlang-core-ast/mir-types';
+  debugPrintMidIRFunction,
+  MidIRFunction,
+  MIR_BOOL_TYPE,
+  MIR_FUNCTION_TYPE,
+  MIR_INT_TYPE,
+} from 'samlang-core-ast/mir-nodes';
 
 it('estimateFunctionInlineCost test', () => {
   expect(
