+++ samlang-core-optimization/__tests__/mir-loop-induction-analysis.test.ts
@@ -23,8 +23,9 @@ import {
   MIR_BREAK,
   MIR_WHILE,
   MIR_CAST,
-} from 'samlang-core-ast/mir-expressions';
-import { MIR_BOOL_TYPE, MIR_INT_TYPE } from 'samlang-core-ast/mir-types';
+  MIR_BOOL_TYPE,
+  MIR_INT_TYPE,
+} from 'samlang-core-ast/mir-nodes';
 import { zip } from 'samlang-core-utils';
 
 const VARIABLE_I = MIR_VARIABLE('i', MIR_INT_TYPE);
