+++ samlang-core-optimization/mir-loop-induction-variable-elimination.ts
@@ -4,14 +4,14 @@ import {
 } from './mir-loop-induction-analysis';
 import type OptimizationResourceAllocator from './optimization-resource-allocator';
 
+import createMidIRFlexibleOrderOperatorNode from 'samlang-core-ast/mir-flexible-op';
 import {
   MidIRExpression,
   MidIRStatement,
   MIR_VARIABLE,
   MIR_BINARY,
-} from 'samlang-core-ast/mir-expressions';
-import createMidIRFlexibleOrderOperatorNode from 'samlang-core-ast/mir-flexible-op';
-import { MIR_INT_TYPE } from 'samlang-core-ast/mir-types';
+  MIR_INT_TYPE,
+} from 'samlang-core-ast/mir-nodes';
 import { checkNotNull } from 'samlang-core-utils';
 
 const midIRLoopInductionVariableEliminationOptimization = (
