+++ samlang-core-printer/index.ts
@@ -2,7 +2,7 @@ import { createPrettierDocumentFromMidIRModule } from './printer-js';
 import { prettyPrintAccordingToPrettierAlgorithm } from './printer-prettier-core';
 import createPrettierDocumentForSamlangModule from './printer-source-level';
 
-import type { MidIRModule } from 'samlang-core-ast/mir-toplevel';
+import type { MidIRModule } from 'samlang-core-ast/mir-nodes';
 import type { SamlangModule } from 'samlang-core-ast/samlang-toplevel';
 
 export const prettyPrintSamlangModule = (
